+++ pages/api/preview.js
@@ -1,35 +1,60 @@
-import { globalDrupalStateAuthStores } from '../../lib/stores';
+import { globalDrupalStateStores } from '../../lib/stores';
 
 const preview = async (req, res) => {
+	const { secret, slug, objectName } = req.query;
 	// Check the secret and next parameters
 	// This secret should only be known to this API route and the CMS
-	if (req.query.secret !== process.env.PREVIEW_SECRET || !req.query.slug) {
-		return res.redirect('/500');
+	if (secret !== process.env.PREVIEW_SECRET) {
+		return res.redirect(
+			`/preview-error/?error=${encodeURIComponent(
+				'Preview secret does not match',
+			)}&message=${encodeURIComponent(
+				'Check that your PREVIEW_SECRET environment variable matches the preview secret generated when creating the preview site in Drupal.',
+			)}`,
+		);
+	}
+
+	if (!slug) {
+		return res.redirect(
+			`/preview-error/?error=${encodeURIComponent(
+				'Requested preview path does not exist',
+			)}&message=${encodeURIComponent('Make sure the content is published')}`,
+		);
 	}
 
 	// returns the store that matches the locale found in the requested url
 	// or the only store if using a monolingual backend
-	const [store] = globalDrupalStateAuthStores.filter(({ defaultLocale }) => {
+	const [store] = globalDrupalStateStores.filter(({ defaultLocale }) => {
 		const regex = new RegExp(`/${defaultLocale}/`);
 		return defaultLocale ? regex.test(req.url) : true;
 	});
-	const objectName = req.query.objectName;
 	// verify the content exists
 	let content;
 	try {
 		content = await store.getObjectByPath({
-			objectName: objectName,
-			path: req.query.slug,
+			objectName,
+			path: slug,
 		});
 	} catch (error) {
 		process.env.DEBUG_MODE &&
-			console.error('Error verifying preview content: ', error);
-		return res.redirect('/500');
+			console.error(
+				'Error verifying preview content in pages/api/preview: ',
+				error,
+			);
+		return res.redirect(
+			`/preview-error/?error=${encodeURIComponent(
+				'Could not verify preview content',
+			)}&message=${encodeURIComponent(error.message)}`,
+		);
 	}
 
 	// If the content doesn't exist prevent preview mode from being enabled
 	if (!content) {
-		return res.redirect('/500');
+		return res.redirect(
+			`/preview-error/?error=${encodeURIComponent(
+				'Requested preview content does not exist',
+			)}&message=${encodeURIComponent('Make sure the content is published')}`,
+		);
 	}
 
 	// Enable Preview Mode by setting a cookie
@@ -49,7 +74,7 @@ const preview = async (req, res) => {
 	}
 
 	// Redirect to the path from the fetched content
-	res.redirect(`${content.path.alias}?timestamp=${Date.now()}`);
+	res.redirect(`${content.path.alias || slug}?timestamp=${Date.now()}`);
 };
 
 export default preview;