+++ pages/pages/[...alias].jsx
@@ -3,7 +3,6 @@ import { isMultiLanguage } from '../../lib/isMultiLanguage';
 import { getPreview } from '../../lib/getPreview';
 import {
 	getCurrentLocaleStore,
-	globalDrupalStateAuthStores,
 	globalDrupalStateStores,
 } from '../../lib/stores';
 
@@ -21,12 +20,12 @@ export default function PageTemplate({ page, footerMenu, hrefLang, preview }) {
 			<article className="prose lg:prose-xl mt-10 mx-auto">
 				<h1>{page.title}</h1>
 
-				<Link passHref href="/pages">
-					<a className="font-normal">Pages &rarr;</a>
+				<Link passHref href="/pages" className="font-normal">
+					Pages &rarr;
 				</Link>
 
 				<div className="mt-12 max-w-lg mx-auto lg:grid-cols-3 lg:max-w-screen-lg">
-					<div dangerouslySetInnerHTML={{ __html: page.body.value }} />
+					<div dangerouslySetInnerHTML={{ __html: page.body.processed }} />
 				</div>
 			</article>
 		</Layout>
@@ -37,10 +36,7 @@ export async function getServerSideProps(context) {
 	const { locales, locale } = context;
 	const multiLanguage = isMultiLanguage(context.locales);
 	const lang = context.preview ? context.previewData.previewLang : locale;
-	const store = getCurrentLocaleStore(
-		lang,
-		context.preview ? globalDrupalStateAuthStores : globalDrupalStateStores,
-	);
+	const store = getCurrentLocaleStore(lang, globalDrupalStateStores);
 
 	// handle nested alias like /pages/featured
 	const alias = `${context.params.alias
@@ -49,6 +45,15 @@ export async function getServerSideProps(context) {
 
 	const previewParams =
 		context.preview && (await getPreview(context, 'node--page'));
+
+	if (previewParams?.error) {
+		return {
+			redirect: {
+				destination: previewParams.redirect,
+				permanent: false,
+			},
+		};
+	}
 
 	let page;
 	try {
@@ -59,6 +64,7 @@ export async function getServerSideProps(context) {
 			params: context.preview && previewParams,
 			refresh: true,
 			res: context.res,
+			anon: context.preview ? false : true,
 		});
 	} catch (error) {
 		// retry the fetch with `/pages` prefix
@@ -69,6 +75,7 @@ export async function getServerSideProps(context) {
 			params: context.preview && previewParams,
 			refresh: true,
 			res: context.res,
+			anon: context.preview ? false : true,
 		});
 	}
 
@@ -76,6 +83,7 @@ export async function getServerSideProps(context) {
 		objectName: 'menu_items--main',
 		refresh: true,
 		res: context.res,
+		anon: true,
 	});
 
 	const origin = process.env.NEXT_PUBLIC_FRONTEND_URL;
@@ -83,7 +91,7 @@ export async function getServerSideProps(context) {
 	const paths = locales.map(async (locale) => {
 		const storeByLocales = getCurrentLocaleStore(
 			locale,
-			context.preview ? globalDrupalStateAuthStores : globalDrupalStateStores,
+			globalDrupalStateStores,
 		);
 		const { path } = await storeByLocales.getObject({
 			objectName: 'node--page',
@@ -91,6 +99,7 @@ export async function getServerSideProps(context) {
 			params: context.preview && previewParams,
 			refresh: true,
 			res: context.res,
+			anon: context.preview ? false : true,
 		});
 		return path;
 	});