+++ pages/pages/index.jsx
@@ -38,8 +38,9 @@ export default function PageListTemplate({
 									href={`${
 										multiLanguage ? `/${path?.langcode || locale}` : ''
 									}${path.alias}`}
+									className="font-normal underline"
 								>
-									<a className="font-normal underline">Read more →</a>
+									Read more →
 								</Link>
 							</li>
 						))
@@ -73,12 +74,14 @@ export async function getServerSideProps(context) {
 			refresh: true,
 			res: context.res,
 			params: 'fields[node--page]=id,title,body,path',
+			anon: true,
 		});
 
 		const footerMenu = await store.getObject({
 			objectName: 'menu_items--main',
 			refresh: true,
 			res: context.res,
+			anon: true,
 		});
 
 		if (!pages) {