+++ CHANGELOG.md
@@ -1,5 +1,131 @@
 # @pantheon-systems/next-drupal-starter
 
+## 4.0.0
+
+### Major Changes
+
+- 0af5b4f: React 18 and Next 13 updates
+
+### Patch Changes
+
+- Updated dependencies [41221dc]
+  - @pantheon-systems/nextjs-kit@1.7.0
+
+## 4.0.0-canary.0
+
+### Major Changes
+
+- 0af5b4f: React 18 and Next 13 updates
+
+### Patch Changes
+
+- Updated dependencies [41221dc]
+  - @pantheon-systems/nextjs-kit@1.7.0-canary.0
+
+## 3.8.0
+
+### Minor Changes
+
+- 4d36cc3: Set IS_LIVE_ENVIRONMENT Environment Variable
+
+### Patch Changes
+
+- 1c63118: Fix 404 on preview when content is missing url alias
+- 07159e7: Support Inline Images In Drupal Body Field
+- d85500a: Add an error message to `getPreview` in case `fetchedPreviewData` is
+  a draft and has no `id`
+- Updated dependencies [4188395]
+  - @pantheon-systems/drupal-kit@4.1.0
+
+## 3.8.0-canary.1
+
+### Patch Changes
+
+- d85500a: Add an error message to `getPreview` in case `fetchedPreviewData` is
+  a draft and has no `id`
+- Updated dependencies [4188395]
+  - @pantheon-systems/drupal-kit@4.1.0-canary.0
+
+## 3.8.0-canary.0
+
+### Minor Changes
+
+- 4d36cc3: Set IS_LIVE_ENVIRONMENT Environment Variable
+
+### Patch Changes
+
+- 1c63118: Fix 404 on preview when content is missing url alias
+- 07159e7: Support Inline Images In Drupal Body Field
+
+## 3.7.0
+
+### Minor Changes
+
+- 56898ad: Return error from getPreview catch, and redirect on that error
+- c32b6cb: Create Preview Error Route
+- b5fae1d: Removed preview-ribbon component and use the one from next-kit
+
+### Patch Changes
+
+- Updated dependencies [276c9d2]
+- Updated dependencies [b5fae1d]
+- Updated dependencies [a529abf]
+  - @pantheon-systems/nextjs-kit@1.6.0
+
+## 3.7.0-canary.2
+
+### Patch Changes
+
+- Updated dependencies [276c9d2]
+  - @pantheon-systems/nextjs-kit@1.6.0-canary.2
+
+## 3.7.0-canary.1
+
+### Minor Changes
+
+- 56898ad: Return error from getPreview catch, and redirect on that error
+
+### Patch Changes
+
+- Updated dependencies [a529abf]
+  - @pantheon-systems/nextjs-kit@1.6.0-canary.1
+
+## 3.7.0-canary.0
+
+### Minor Changes
+
+- c32b6cb: Create Preview Error Route
+- b5fae1d: Removed preview-ribbon component and use the one from next-kit
+
+### Patch Changes
+
+- Updated dependencies [b5fae1d]
+  - @pantheon-systems/nextjs-kit@1.6.0-canary.0
+
+## 3.6.0
+
+### Minor Changes
+
+- 896d70a: Refactor stores to use DS auth/anon request flow feature
+
+### Patch Changes
+
+- Updated dependencies [a5c4275]
+- Updated dependencies [cd286c6]
+  - @pantheon-systems/drupal-kit@4.0.0
+
+## 3.6.0-canary.0
+
+### Minor Changes
+
+- 896d70a: Refactor stores to use DS auth/anon request flow feature
+
+### Patch Changes
+
+- Updated dependencies [a5c4275]
+- Updated dependencies [cd286c6]
+  - @pantheon-systems/drupal-kit@4.0.0-canary.0
+
 ## 3.5.0
 
 ### Minor Changes